Change usage of ParametersConverter
The RequestParser now converts the parameters right away so that they are
ready for usage with the server (i.e. respects the parameters->server
section of the configuration). I also removed the ParametersConverter from
all constructors, so that it becomes entirely an implementation detail and
cannot be switched out with Pimple. That is, the two classes in
Core\Client instantiate a converter themselves.